desktop-organizer/requirements.txt
# Core
PyYAML>=6.0
